Simplify and improve the Topic Operator performance test.
This change builds on the work that Maros did, removing the reporting stuff and making the test faster.
The test runs 4 batches of topic events (10, 100, 500, 100) and it takes a 5m40s on my machines.
It can be used to evaluate the impact on performance when reviewing Topic Operator changes.
On the maintenance side, it should not be more complicated than a regular system test.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>